Make dice common cookbook Chef compliant
Chef does not like to have dashes in cookbook and recipe names. In
order to be future proof, we renamed these files now when there is
manageable amount of packages that depend on this common cookbook.

While doing the cleanup, we also took the opportunity and removed all
errors and warnings that rubocop reported. This should make it easier
to keep reciper Chef compliant in the future.

Change-Id: Ie5050fe6f34c64c8f64923dd3c6ab3fdffea3b4f