Add standalone Apache Zookeeper cookbook
Existing cookbook for zookeeper is getting old, but we cannot update
it because we made profound modifications of upstream cookbook that
make it almost impossible to merge changes cleanly.

Instead, we opted for minimalistic rewrite of the cookbook that should
be drop-in replacement for currently used zookeeper recipes. Zookeeper
is installed from upstream-provided builds and do not rely on any 3rd
party vendor to update it's packages.

Yet another benefit of this move is test suite: we can now run test
that should make it a bit safer to update zookeeper without breaking
everything.

Change-Id: I8f06a9051c89c0b2dde934eaacd4669ced33bc24