Update Apache Spark cookbook
In order to be able to write sane spark job submitter, we needed to do
some cleanup. Most important change is installation of command
wrappers that submitter can use to perform requested operation. These
wrappers are needed because any settings that are used to configure
spark installation are lost when Chef run terminates.

Second major thing addition of version selection. We now officially
support Apache Spark 1.6.2 and 2.0.1 (they can even be installed in
parallel with no problems). Wrappers are installed for all versions
(versioned commands are in the form of <command>-<version>, for
example spark-submit-1.6). By default, we also install non-suffixed
alias for user's convenience.

Change-Id: I182b098ba5c96e6096f69bad8425d1085d630e3b