Add checksum to dmon agent package
Checksum has been added in order to prevent working on broken file
(for example, on partially downloaded file).

While implementing this change, we discovered that file being
downloaded is not really gzipped, but is plain old tarball. In order
to accommodate this fact, download code has been remodeled a bit.

Change-Id: I4a89a7bb262423fd180e1d9213bfa8aca734e100