Install cqlsh wrapper
In order to make it easier to test things on installed Cassandra
cluster, we install cqlsh wrapper that properly connects to Cassandra
without any user input.

This will also come in handy when we implement Cassandra preparation
TOSCA type, since most of the stuff will probably rely on ability to
execute CQL commands.

Change-Id: Iff8daec1b1d84a5ad82c25ea8cebb1e23f5852e9