Fix java versions and flavor in recipes
Our kitchen files were a bit inconsistent when it comes to specifying
java versions. AS things stand now, all of the TOSCA types we defined
use java version 8 and this is what we set here.

If we need to support multiple versions of java, additional suites
with separate java settings will need to be added.

Change-Id: Ia36f98db869b1aafb065ad36e2c5eb96a37e366e