Add default collectd types to main configuration
Collectd is a bit strange when it comes to TypesDB directive. If there
is no instance of this directive present in configuration, default
types are used. If we add custom types, default types are ignored if
not explicitly requested, which is not what we want.

To remedy this situation we added default types to the main collectd
configuration file, which makes sure default types are always
available, not matter what any included files do to type database.

Change-Id: I574f5a1d270f4679e6a046205184e09765178045