Generate DDS certificate on the spot if needed
Since not all platforms support floating IPs, we need to resort to
generating certificates during installation. This way we can set the
proper IP address in SAN field of the configuration.

We did not remove support for user supplied certificates, we only
added an options of not supplying them and still getting working
service over https.

Change-Id: I794396ed17762ee2aec0cc886705f64f1efcd715