Allow blueprint to set node role
Default recipe now checks if any roles have been set from within the
blueprint. This allows blueprint authors to set custom roles for their
nodes (for example, when using a script runner node that can install
almost anything on VM).

Note that we do not perform any validation of roles. If the role is
not supported by DMon, behavior of the default recipe is undefined and
depends on what response will come from DMon when setting a role.

Change-Id: I7a25e2775af252dcaf26c96e94a9b8df99f5af4f