Allow running dns_redirect recipe multiple times
At the end of the DNS settings update, we make them immutable in order
to protect them from system tools that would change them. But this
immutability can cause troubles when we run the same recipe again on
this same machine, since immutability prevents chef to do its work.

To fix this, we simply make resolv.conf file mutable for the duration
of the chef recipe run.

Change-Id: Iaad61f21401e16f5587c7be60adb3e7ea21577f1