Only set host name on first run
Previously, each run of the cookbook produced different host name if
the node names were different (which were most of the time). This
caused some troubles with different services, since VM could be
registered to consul multiple times.

This fix adds marker to hosts file that indicates that we already set
the host name and FQDN. If this marker is encountered, we simply skip
the recipe.

Change-Id: Ia52e17a5969ae696ff2a1b36270c6711ec77af22