Install proper database for service data storage
Up until now, default recipes installed DICE deployment service with
development SQLite database storage, which was not optimal to put it
mildly.

When service was put to good use and concurrency of database access
increased, we started to get exceptions because database was locked
most of the time. In order to avoid this, we now install PostgreSQL as
a data storage for deployment service.

Change-Id: Iee76300aa546f55d4f81a41dfb47c7d5d05e8946