Add installing `dmon-agent` as service for Storm
DICE Monitoring Platform requires custom access to the log files
generated by the Storm services such as the supervisor. The
`dmon-agent` provides a REST call for this functionality, thus
the deployer needs to install the agent as a service and start
it when installing Storm.

This commit adds a recipe `start` for starting the `dmon-agent`
service. It also adds steps in the `default` recipe to install
the Upstart job for `dmon-agent`.

Change-Id: I9614d0b2604ef9b675110726314fe62758d869f3