Add configuration to DMon agent upstart job
Since DMon agent needs to know thing or two about the installation
that it is monitoring, we need to have a more general way of
customizing the environment that DMon agent is being run in.

We solved this by adding a configuration folder for upstart job that
can contain files that set environment variables that main DMon agent
process will see. This way, any service that needs to inform DMon
agent about its configuration can do so by simply placing a file in
proper folder and restarting the service.

Change-Id: I4318f23ca40cd19ee56e15ca88e7fb378837def8