Update DMon agent package
Older DMon agent was not able to cope with Apache Storm 1.0.x logs.
Newly released 0.2.6x version fixed this issue.

New version of DMon agent needs to be explicitly informed about the
version of Apache Storm that is running on the node, since if the
STORM_VERSION environment variable is not set, agent will resort to
fetching logs for Apache Storm 0.9.x.

Change-Id: If8141381b05fc5ab8f63eae46a56e3932126c59d