Fix MongoDB bind settings
When we created the mongo configuration file, we always added loopback
address to the bind addresses, which causes mongo to stop responding
if the other bind address is 0.0.0.0. Note that mongo daemon is still
running and does not abort the start o this error, which is why we
missed this situation in the first place.

Fix for this situation is quite simple: if we bind to 0.0.0.0, no
other IP should be present in net.bindIp configuration.

Change-Id: I46c17603978c9be7e76d07144f2935f3fd74ffb9