Disable test on DICE Deployment Service CLI tool
Command-line tool for interacting with DICE Deployment Service has
been installed globally in order to make it easier to use. With the
release of pip 10, this installation method broke down because pip
(rightfully) refuses to replace system packages.

Until this is fixed, test is disabled and cookbook should not be used
in production.

Change-Id: I0a47c7bbcb1718410a0e15612dc3b4eaddf35808